feat: roles and permissions, change password, delete account
Resolves #254.
Resolves #580.
Resolves #581.
