+++ app/Http/Controllers/InvitationController.php
@@ -33,7 +33,7 @@ class InvitationController extends Controller
 
         flash(__('invitation.create_invitation_succeeded'), 'success');
 
-        return redirect(localized_route('users.edit-roles-and-permissions'));
+        return redirect(localized_route('settings.edit-roles-and-permissions'));
     }
 
     /**
@@ -74,6 +74,6 @@ class InvitationController extends Controller
 
         flash(__('invitation.cancel_invitation_succeeded'), 'success');
 
-        return redirect(localized_route('users.edit-roles-and-permissions'));
+        return redirect(localized_route('settings.edit-roles-and-permissions'));
     }
 }
