+++ app/Http/Controllers/UserController.php
@@ -11,7 +11,6 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Auth;
 use function localized_route;
-use Spatie\LaravelOptions\Options;
 
 class UserController extends Controller
 {
@@ -147,68 +146,6 @@ class UserController extends Controller
     }
 
     /**
-     * Show the roles and permissions edit view for the logged-in user.
-     *
-     * @return View
-     */
-    public function editRolesAndPermissions(): View
-    {
-        $user = Auth::user();
-        $roles = [];
-
-        foreach (config('hearth.organizations.roles') as $role) {
-            $roles[$role] = __('roles.'.$role);
-        }
-
-        return view('users.roles-and-permissions', [
-            'user' => $user,
-            'roles' => Options::forArray($roles)->toArray(),
-            'regulatedOrganization' => $user->context === 'regulated-organization' ? $user->regulatedOrganization : null,
-        ]);
-    }
-
-    /**
-     * Show the roles and permissions edit view for the logged-in user.
-     *
-     * @return View|RedirectResponse
-     */
-    public function inviteToInvitationable(): View|RedirectResponse
-    {
-        $currentUser = Auth::user();
-        $invitationable = match ($currentUser->context) {
-            'organization' => $currentUser->organization ?? null,
-            'regulated-organization' => $currentUser->regulatedOrganization ?? null,
-            default => null,
-        };
-
-        if ($invitationable) {
-            $roles = [];
-
-            foreach (config('hearth.organizations.roles') as $role) {
-                $roles[$role] = __('roles.'.$role);
-            }
-
-            return view('users.roles-and-permissions.invite', [
-                'user' => $currentUser,
-                'invitationable' => $invitationable,
-                'roles' => Options::forArray($roles)->toArray(),
-            ]);
-        }
-
-        return redirect(localized_route('users.edit-roles-and-permissions'));
-    }
-
-    /**
-     * Show the password and security admin view for the logged-in user.
-     *
-     * @return View
-     */
-    public function admin(): View
-    {
-        return view('users.admin', ['user' => Auth::user()]);
-    }
-
-    /**
      * Show the "my projects" view for the logged-in user.
      *
      * @return RedirectResponse|View
@@ -225,16 +162,6 @@ class UserController extends Controller
         }
 
         return redirect(localized_route('dashboard'));
-    }
-
-    /**
-     * Show the account deletion view for the logged-in user.
-     *
-     * @return View
-     */
-    public function delete(): View
-    {
-        return view('users.delete', ['user' => Auth::user()]);
     }
 
     /**
