+++ resources/views/settings/delete-account.blade.php
