+++ resources/views/settings/roles-and-permissions/invite.blade.php
@@ -3,7 +3,7 @@
     <x-slot name="header">
         <ol class="breadcrumbs" role="list">
             <li><a href="{{ localized_route('settings.show') }}">{{ __('Settings') }}</a></li>
-            <li><a href="{{ localized_route('users.edit-roles-and-permissions') }}">{{ __('Roles and permissions') }}</a></li>
+            <li><a href="{{ localized_route('settings.edit-roles-and-permissions') }}">{{ __('Roles and permissions') }}</a></li>
         </ol>
         <h1>
             {{ __('Invite new member') }}
