+++ tests/Feature/PasswordConfirmationTest.php
@@ -1,44 +1,32 @@
 <?php
 
-namespace Tests\Feature;
-
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
 
-class PasswordConfirmationTest extends TestCase
-{
-    use RefreshDatabase;
+test('confirm password screen can be rendered', function () {
+    $user = User::factory()->create();
 
-    public function test_confirm_password_screen_can_be_rendered()
-    {
-        $user = User::factory()->create();
+    $response = $this->actingAs($user)->get(localized_route('password.confirm'));
 
-        $response = $this->actingAs($user)->get(localized_route('password.confirm'));
+    $response->assertOk();
+});
 
-        $response->assertOk();
-    }
+test('password can be confirmed', function () {
+    $user = User::factory()->create();
 
-    public function test_password_can_be_confirmed()
-    {
-        $user = User::factory()->create();
+    $response = $this->actingAs($user)->post(localized_route('password.confirm'), [
+        'password' => 'password',
+    ]);
 
-        $response = $this->actingAs($user)->post(localized_route('password.confirm'), [
-            'password' => 'password',
-        ]);
+    $response->assertRedirect();
+    $response->assertSessionHasNoErrors();
+});
 
-        $response->assertRedirect();
-        $response->assertSessionHasNoErrors();
-    }
+test('password is not confirmed with invalid password', function () {
+    $user = User::factory()->create();
 
-    public function test_password_is_not_confirmed_with_invalid_password()
-    {
-        $user = User::factory()->create();
+    $response = $this->actingAs($user)->post(localized_route('password.confirm'), [
+        'password' => 'wrong-password',
+    ]);
 
-        $response = $this->actingAs($user)->post(localized_route('password.confirm'), [
-            'password' => 'wrong-password',
-        ]);
-
-        $response->assertSessionHasErrors();
-    }
-}
+    $response->assertSessionHasErrors();
+});
